Extracted wohngeld functions to extra module.
